rename namespace back to DSCore.IO
