+++ src/Libraries/CoreNodes/DSCoreNodes.Migrations.xml
@@ -26,7 +26,7 @@
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.CSV.WriteToFile</oldName>
-    <newName>DSCore.Data.ExportCSV</newName>
+    <newName>DSCore.IO.Data.ExportCSV</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.Display.ByGeometryColor</oldName>
@@ -68,27 +68,27 @@
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.HasExtension</oldName>
-    <newName>DSCore.File.FileHasExtension</newName>
+    <newName>DSCore.IO.File.FileHasExtension</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.ChangeExtension</oldName>
-    <newName>DSCore.File.ChangePathExtension</newName>
+    <newName>DSCore.IO.File.ChangePathExtension</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.DirectoryName</oldName>
-    <newName>DSCore.File.DirectoryName</newName>
+    <newName>DSCore.IO.File.DirectoryName</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.Combine</oldName>
-    <newName>DSCore.File.CombinePath</newName>
+    <newName>DSCore.IO.File.CombinePath</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.FileName</oldName>
-    <newName>DSCore.File.FileName</newName>
+    <newName>DSCore.IO.File.FileName</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.FilePath.Extension</oldName>
-    <newName>DSCore.File.FileExtension</newName>
+    <newName>DSCore.IO.File.FileExtension</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.Formula.Evaluate</oldName>
