+++ test/Libraries/CoreNodesTests/FileTests.cs
@@ -12,7 +12,6 @@ using NUnit.Framework;
 using Color = DSCore.Color;
 using Directory = System.IO.Directory;
 using File = DSCore.IO.File;
-using FilePath = DSCore.File;
 using Image = DSCore.IO.Image;
 
 namespace Dynamo.Tests
@@ -29,28 +28,28 @@ namespace Dynamo.Tests
         [Test, Category("UnitTests")]
         public void SimpleWrappers()
         {
-            Assert.AreEqual(Path.Combine("test"), FilePath.CombinePath("test"));
-            Assert.AreEqual(Path.Combine("test", "1"), FilePath.CombinePath("test", "1"));
-            Assert.AreEqual(Path.Combine("test/", @"1\"), FilePath.CombinePath("test/", @"1\"));
+            Assert.AreEqual(Path.Combine("test"), File.CombinePath("test"));
+            Assert.AreEqual(Path.Combine("test", "1"), File.CombinePath("test", "1"));
+            Assert.AreEqual(Path.Combine("test/", @"1\"), File.CombinePath("test/", @"1\"));
 
             const string aFilePath = @"hello\there.txt";
             const string aFileName = "hello";
 
-            Assert.AreEqual(Path.GetExtension(aFilePath), FilePath.FileExtension(aFilePath));
-            Assert.AreEqual(Path.GetExtension(aFileName), FilePath.FileExtension(aFileName));
+            Assert.AreEqual(Path.GetExtension(aFilePath), File.FileExtension(aFilePath));
+            Assert.AreEqual(Path.GetExtension(aFileName), File.FileExtension(aFileName));
 
             Assert.AreEqual(
                 Path.ChangeExtension(aFilePath, ".png"),
-                FilePath.ChangePathExtension(aFilePath, ".png"));
+                File.ChangePathExtension(aFilePath, ".png"));
             Assert.AreEqual(
                 Path.ChangeExtension(aFileName, ".txt"),
-                FilePath.ChangePathExtension(aFileName, ".txt"));
+                File.ChangePathExtension(aFileName, ".txt"));
 
-            Assert.AreEqual(Path.GetDirectoryName(aFilePath), FilePath.DirectoryName(aFilePath));
-            Assert.AreEqual(Path.GetDirectoryName(aFileName), FilePath.DirectoryName(aFileName));
+            Assert.AreEqual(Path.GetDirectoryName(aFilePath), File.DirectoryName(aFilePath));
+            Assert.AreEqual(Path.GetDirectoryName(aFileName), File.DirectoryName(aFileName));
 
-            Assert.AreEqual(Path.HasExtension(aFilePath), FilePath.FileHasExtension(aFilePath));
-            Assert.AreEqual(Path.HasExtension(aFileName), FilePath.FileHasExtension(aFileName));
+            Assert.AreEqual(Path.HasExtension(aFilePath), File.FileHasExtension(aFilePath));
+            Assert.AreEqual(Path.HasExtension(aFileName), File.FileHasExtension(aFileName));
         }
 
         [Test, Category("UnitTests")]
@@ -58,10 +57,10 @@ namespace Dynamo.Tests
         {
             const string aFilePath = @"hello\there.txt";
 
-            Assert.AreEqual(Path.GetFileName(aFilePath), FilePath.FileName(aFilePath));
+            Assert.AreEqual(Path.GetFileName(aFilePath), File.FileName(aFilePath));
             Assert.AreEqual(
                 Path.GetFileNameWithoutExtension(aFilePath),
-                FilePath.FileName(aFilePath, withExtension: false));
+                File.FileName(aFilePath, withExtension: false));
         }
         #endregion
 
@@ -237,14 +236,14 @@ namespace Dynamo.Tests
             var tmpSrc = GetNewFileNameOnTempPath("");
             Directory.CreateDirectory(tmpSrc);
             const string fileName = @"temp.txt";
-            File.WriteText(FilePath.CombinePath(tmpSrc, fileName), "test");
+            File.WriteText(File.CombinePath(tmpSrc, fileName), "test");
 
             var tmpDest = GetNewFileNameOnTempPath("");
             DSCore.IO.Directory.Move(tmpSrc, tmpDest);
             Assert.IsFalse(DSCore.IO.Directory.Exists(tmpSrc));
             Assert.IsTrue(DSCore.IO.Directory.Exists(tmpDest));
 
-            var destFileName = FilePath.CombinePath(tmpDest, fileName);
+            var destFileName = File.CombinePath(tmpDest, fileName);
             Assert.IsTrue(File.Exists(destFileName));
             Assert.AreEqual("test", File.ReadText(File.FromPath(destFileName)));
         }
@@ -255,14 +254,14 @@ namespace Dynamo.Tests
             var tmpSrc = GetNewFileNameOnTempPath("");
             var tmpSrcInfo = DSCore.IO.Directory.FromPath(tmpSrc);
             const string fileName = @"temp.txt";
-            File.WriteText(FilePath.CombinePath(tmpSrc, fileName), "test");
+            File.WriteText(File.CombinePath(tmpSrc, fileName), "test");
 
             var tmpDest = GetNewFileNameOnTempPath("");
             DSCore.IO.Directory.Copy(tmpSrcInfo, tmpDest);
             Assert.IsTrue(DSCore.IO.Directory.Exists(tmpSrc));
             Assert.IsTrue(DSCore.IO.Directory.Exists(tmpDest));
 
-            var destFileName = FilePath.CombinePath(tmpDest, fileName);
+            var destFileName = File.CombinePath(tmpDest, fileName);
             Assert.IsTrue(File.Exists(destFileName));
             Assert.AreEqual("test", File.ReadText(File.FromPath(destFileName)));
         }
@@ -273,7 +272,7 @@ namespace Dynamo.Tests
             var tmpSrc = GetNewFileNameOnTempPath("");
             Directory.CreateDirectory(tmpSrc);
             const string fileName = @"temp.txt";
-            File.WriteText(FilePath.CombinePath(tmpSrc, fileName), "test");
+            File.WriteText(File.CombinePath(tmpSrc, fileName), "test");
 
             Assert.Throws<IOException>(() => DSCore.IO.Directory.Delete(tmpSrc));
             DSCore.IO.Directory.Delete(tmpSrc, recursive: true);
@@ -291,11 +290,11 @@ namespace Dynamo.Tests
             var tmpSrc = GetNewFileNameOnTempPath("");
             var tmpSrcInfo = DSCore.IO.Directory.FromPath(tmpSrc);
             const string fileName = @"temp.txt";
-            var newFile = FilePath.CombinePath(tmpSrc, fileName);
+            var newFile = File.CombinePath(tmpSrc, fileName);
             File.WriteText(newFile, "test");
 
             const string dirName = @"subDir";
-            var newDir = FilePath.CombinePath(tmpSrc, dirName);
+            var newDir = File.CombinePath(tmpSrc, dirName);
             Directory.CreateDirectory(newDir);
 
             var contents = DSCore.IO.Directory.Contents(tmpSrcInfo);
